+++ .babelrc
@@ -1,0 +1,7 @@
+{
+  "presets": [
+    "es2015",
+    "stage-0",
+    "react"
+  ]
+}
