+++ src/components/Main.js
@@ -1,22 +1,24 @@
-require('normalize.css/normalize.css');
-require('styles/App.css');
+import 'normalize.css/normalize.css';
+import 'styles/App.css';
 
 import React from 'react';
-import Recipes from '../components/Recipes';
+import Recipes from '../models/Recipes';
 
 class AppComponent extends React.Component {
   render() {
-    let recipesRepo = new Recipes
-    let recipes = recipesRepo.search()
+    const recipesRepo = new Recipes;
+    const recipes = recipesRepo.search();
     return (
       <div className="index">
         <h1 className="title">Recipies</h1>
         <div className="search-box"><input type="text" value="Search here..." /></div>
         <div className="recipes-list">
         {recipes.map((recipe) => {
-        	return <div className="recipe-wrap">
-        	  <a href="#" title={recipe.description}><span className="recipe-name">{recipe.name}</span></a>
-        	</div>
+          return (
+            <div className="recipe-wrap">
+              <a href="#" title={recipe.description}><span className="recipe-name">{recipe.name}</span></a>
+            </div>
+          );
         })}
         </div>
         <div className="footer">Footer</div>
