--- src/components/Recipes.js
@@ -1,21 +1,0 @@
-class Recipes {
-
-	search() {
-		return [
-		  {name: 'Recipe 01', description: 'Random stuff 01'},
-		  {name: 'Recipe 02', description: 'Random stuff 02'},
-		  {name: 'Recipe 03', description: 'Random stuff 03'},
-		  {name: 'Recipe 04', description: 'Random stuff 04'},
-		  {name: 'Recipe 05', description: 'Random stuff 05'},
-		  {name: 'Recipe 06', description: 'Random stuff 06'},
-		  {name: 'Recipe 07', description: 'Random stuff 07'},
-		  {name: 'Recipe 08', description: 'Random stuff 08'},
-		  {name: 'Recipe 09', description: 'Random stuff 09'},
-		  {name: 'Recipe 10', description: 'Random stuff 10'},
-		  {name: 'Recipe 11', description: 'Random stuff 11'},
-		  {name: 'Recipe 12', description: 'Random stuff 12'}
-		];
-	}
-}
-
-export default Recipes
